moved gruntfile functions to utility
