+++ .env_example
@@ -1,16 +1,11 @@
-STEAM_API_KEY=REPLACE_WITH_KEY
 SESSION_SECRET=somesecretvalue
-ROOT_URL=http://yourdomain.com
+ROOT_URL=http://localhost:5000
 START_SEQ_NUM=AUTO
 RETRIEVER_HOST=http://localhost:5200,http://retrieverhost2.com
-RECAPTCHA_PUBLIC_KEY=REPLACE_WITH_KEY
-RECAPTCHA_SECRET_KEY=REPLACE_WITH_KEY
 KUE_USER=user
 KUE_PASS=pass
 DELETE_REPLAYS=false
-AWS_S3_BUCKET=somebucket
-AWS_ACCESS_KEY_ID=REPLACE_WITH_KEY
-AWS_SECRET_ACCESS_KEY=REPLACE_WITH_KEY
+AWS_S3_BUCKET=optionalbucket
 NODE_ENV=development
 UNTRACK_INTERVAL_DAYS=3
 PORT=5000
@@ -20,3 +15,9 @@ STEAM_USER=user1,user2
 STEAM_PASS=pass1,pass2
 STEAM_GUARD_CODE=""
 RETRIEVER=false
+REMOTE=http://remotesite.com
+STEAM_API_KEY=REPLACE_WITH_KEY
+RECAPTCHA_PUBLIC_KEY=REPLACE_WITH_KEY
+RECAPTCHA_SECRET_KEY=REPLACE_WITH_KEY
+AWS_ACCESS_KEY_ID=REPLACE_WITH_KEY
+AWS_SECRET_ACCESS_KEY=REPLACE_WITH_KEY
