+++ retriever.js
@@ -114,11 +114,10 @@ function reset() {
     lock = false;
 }
 
-module.exports.app = app;
-
 setTimeout(function() {
     process.exit(0);
 }, 1000 * 60 * 60);
+
 var server = app.listen(process.env.PORT || 3000, function() {
     var host = server.address().address;
     var port = server.address().port;
