+++ yasp.js
@@ -1,5 +1,6 @@
 var dotenv = require('dotenv');
 dotenv.load();
+var db = require('./db');
 var express = require('express');
 var session = require('cookie-session');
 var multer = require('multer');
@@ -11,9 +12,6 @@ var utility = require('./utility'),
     moment = require('moment'),
     Recaptcha = require('recaptcha').Recaptcha,
     bodyParser = require('body-parser'),
-    matches = utility.matches,
-    players = utility.players,
-    fs = require('fs'),
     kue = utility.kue,
     redis = utility.redis,
     SteamStrategy = require('passport-steam').Strategy,
@@ -59,7 +57,7 @@ passport.serializeUser(function(user, done) {
     done(null, user.account_id);
 });
 passport.deserializeUser(function(id, done) {
-    players.findAndModify({
+    db.players.findAndModify({
         query: {
             account_id: id
         },
@@ -82,7 +80,7 @@ passport.use(new SteamStrategy({
     insert.account_id = steam32;
     insert.track = 1;
     insert.last_visited = new Date(); // $currentDate only exists in Mongo >= 2.6
-    players.findAndModify({
+    db.players.findAndModify({
         account_id: steam32
     }, {
         $set: insert
@@ -133,7 +131,7 @@ app.param('match_id', function(req, res, next, id) {
     redis.get(id, function(err, reply) {
         if (err || !reply) {
             logger.info("Cache miss for match " + id);
-            matches.findOne({
+            db.matches.findOne({
                 match_id: Number(id)
             }, function(err, match) {
                 if (err || !match) {
@@ -141,6 +139,9 @@ app.param('match_id', function(req, res, next, id) {
                 }
                 else {
                     utility.fillPlayerNames(match.players, function(err) {
+                        if (err) {
+                            return next(err);
+                        }
                         req.match = match;
                         //Add to cache if we have parsed data
                         if (match.parsed_data && process.env.NODE_ENV === "production") {
@@ -175,11 +176,11 @@ app.route('/api/matches').get(function(req, res, next) {
         //options = utility.makeSearch(search, req.query.columns)
         sort = utility.makeSort(req.query.order, req.query.columns);
     }
-    utility.matches.count(options, function(err, count) {
+    db.matches.count(options, function(err, count) {
         if (err) {
             return next(err);
         }
-        utility.matches.find(options, {
+        db.matches.find(options, {
             limit: Number(req.query.length),
             skip: Number(req.query.start),
             sort: sort
@@ -218,38 +219,17 @@ app.route('/matches/:match_id/:info?').get(function(req, res, next) {
                 var hero_id = app.locals.constants.hero_names[key].id;
                 var slot = match.parsed_data.hero_to_slot[hero_id];
                 if (slot) {
-                    var primary = match.players[slot].hero_id
-                    var primary_name = app.locals.constants.heroes[primary].name
-                    var merge = match.parsed_data.heroes[primary_name]
+                    var primary = match.players[slot].hero_id;
+                    var primary_name = app.locals.constants.heroes[primary].name;
+                    var merge = match.parsed_data.heroes[primary_name];
                     if (!match.players[slot].hero_ids) {
-                        match.players[slot].hero_ids = []
+                        match.players[slot].hero_ids = [];
                     }
                     match.players[slot].hero_ids.push(hero_id);
                     if (key !== primary_name) {
-                        mergeObjects(merge, val);
+                        utility.mergeObjects(merge, val);
                     }
                 }
-            }
-        }
-    }
-
-    function mergeObjects(merge, val) {
-        for (var attr in val) {
-            if (val[attr].constructor === Array) {
-                merge[attr] = merge[attr].concat(val[attr])
-            }
-            else if (typeof val[attr] === "object") {
-                mergeObjects(merge[attr], val[attr])
-            }
-            else {
-                //does property exist?
-                if (!merge[attr]) {
-                    merge[attr] = val[attr]
-                }
-                else {
-                    merge[attr] += val[attr]
-                }
-
             }
         }
     }
@@ -260,34 +240,34 @@ app.route('/matches/:match_id/:info?').get(function(req, res, next) {
             var goldDifference = ['Gold'];
             var xpDifference = ['XP'];
             for (var i = 0; i < match.parsed_data.times.length; i++) {
-                var goldtotal = 0
-                var xptotal = 0
+                var goldtotal = 0;
+                var xptotal = 0;
                 match.parsed_data.players.forEach(function(elem, j) {
                     if (match.players[j].player_slot < 64) {
-                        goldtotal += elem.gold[i]
-                        xptotal += elem.xp[i]
+                        goldtotal += elem.gold[i];
+                        xptotal += elem.xp[i];
                     }
                     else {
-                        xptotal -= elem.xp[i]
-                        goldtotal -= elem.gold[i]
+                        xptotal -= elem.xp[i];
+                        goldtotal -= elem.gold[i];
                     }
-                })
-                goldDifference.push(goldtotal)
-                xpDifference.push(xptotal)
+                });
+                goldDifference.push(goldtotal);
+                xpDifference.push(xptotal);
             }
-            var time = ["time"].concat(match.parsed_data.times)
+            var time = ["time"].concat(match.parsed_data.times);
             var data = {
                 difference: [time, goldDifference, xpDifference],
                 gold: [time],
                 xp: [time],
                 lh: [time]
-            }
+            };
             match.parsed_data.players.forEach(function(elem, i) {
-                var hero = app.locals.constants.heroes[match.players[i].hero_id].localized_name
-                data.gold.push([hero].concat(elem.gold))
-                data.xp.push([hero].concat(elem.xp))
-                data.lh.push([hero].concat(elem.lh))
-            })
+                var hero = app.locals.constants.heroes[match.players[i].hero_id].localized_name;
+                data.gold.push([hero].concat(elem.gold));
+                data.xp.push([hero].concat(elem.xp));
+                data.lh.push([hero].concat(elem.lh));
+            });
         }
     }
     res.render(matchPages[info].template, {
@@ -306,22 +286,25 @@ app.route('/players/:account_id/:info?').get(function(req, res, next) {
     if (!playerPages[info]) {
         return next();
     }
-    players.findOne({
+    db.players.findOne({
         account_id: Number(req.params.account_id)
     }, function(err, player) {
-        if (!player) {
-            return next();
+        if (err || !player) {
+            return next(err);
         }
         else {
             utility.getMatches(player.account_id, function(err, matches) {
-                var account_id = player.account_id
-                var counts = {}
-                var heroes = {}
-                player.win = 0
-                player.lose = 0
-                player.games = 0
-                player.teammates = []
-                player.calheatmap = {}
+                if (err) {
+                    return next(err);
+                }
+                var account_id = player.account_id;
+                var counts = {};
+                var heroes = {};
+                player.win = 0;
+                player.lose = 0;
+                player.games = 0;
+                player.teammates = [];
+                player.calheatmap = {};
 
                 for (var i = 0; i < matches.length; i++) {
                     //add start time to data for cal-heatmap
@@ -336,9 +319,9 @@ app.route('/players/:account_id/:info?').get(function(req, res, next) {
                             matches[i].player_win = (playerRadiant === matches[i].radiant_win);
                             matches[i].slot = j;
                             matches[i].player_win ? player.win += 1 : player.lose += 1;
-                            player.games += 1
+                            player.games += 1;
                             if (!heroes[p.hero_id]) {
-                                heroes[p.hero_id] = {}
+                                heroes[p.hero_id] = {};
                                 heroes[p.hero_id]["games"] = 0;
                                 heroes[p.hero_id]["win"] = 0;
                                 heroes[p.hero_id]["lose"] = 0;
@@ -380,6 +363,9 @@ app.route('/players/:account_id/:info?').get(function(req, res, next) {
                 player.matches = matches;
                 player.heroes = heroes;
                 utility.fillPlayerNames(player.teammates, function(err) {
+                    if (err) {
+                        return next(err);
+                    }
                     res.render(playerPages[info].template, {
                         route: info,
                         player: player,
@@ -476,31 +462,31 @@ app.route('/upload')
 app.get('/stats', function(req, res, next) {
     async.parallel({
             matches: function(cb) {
-                utility.matches.count({}, function(err, res) {
+                db.matches.count({}, function(err, res) {
                     cb(err, res);
                 });
             },
             players: function(cb) {
-                utility.players.count({}, function(err, res) {
+                db.players.count({}, function(err, res) {
                     cb(err, res);
                 });
             },
             tracked_players: function(cb) {
-                utility.players.count({
+                db.players.count({
                     track: 1
                 }, function(err, res) {
                     cb(err, res);
                 });
             },
             formerly_tracked_players: function(cb) {
-                utility.players.count({
+                db.players.count({
                     track: 0
                 }, function(err, res) {
                     cb(err, res);
                 });
             },
             matches_last_day: function(cb) {
-                utility.matches.count({
+                db.matches.count({
                     start_time: {
                         $gt: Number(moment().subtract(1, 'day').format('X'))
                     }
@@ -509,28 +495,28 @@ app.get('/stats', function(req, res, next) {
                 });
             },
             queued_matches: function(cb) {
-                utility.matches.count({
+                db.matches.count({
                     parse_status: 0
                 }, function(err, res) {
                     cb(err, res);
                 });
             },
             unavailable_matches: function(cb) {
-                utility.matches.count({
+                db.matches.count({
                     parse_status: 1
                 }, function(err, res) {
                     cb(err, res);
                 });
             },
             parsed_matches: function(cb) {
-                utility.matches.count({
+                db.matches.count({
                     parse_status: 2
                 }, function(err, res) {
                     cb(err, res);
                 });
             },
             uploaded_matches: function(cb) {
-                utility.matches.count({
+                db.matches.count({
                     upload: true
                 }, function(err, res) {
                     cb(err, res);
@@ -565,34 +551,39 @@ app.use(function(req, res) {
     }
 });
 
-var jobs = utility.jobs;
-var trackedPlayers = {};
+module.exports.app = app;
+
+var server = app.listen(process.env.PORT || 3000, function() {
+    var host = server.address().address;
+    var port = server.address().port;
+    console.log('[WEB] listening at http://%s:%s', host, port);
+});
 
 utility.clearActiveJobs('api', function(err) {
     if (err) {
-        logger.info(err);
+        utility.logger.info(err);
     }
-    jobs.process('api', processApiReq);
+    utility.jobs.process('api', utility.processApiReq);
 });
 
 utility.clearActiveJobs('upload', function(err) {
     if (err) {
-        logger.info(err);
+        utility.logger.info(err);
     }
-    jobs.process('upload', processUpload);
+    utility.jobs.process('upload', utility.processUpload);
 });
 
 if (process.env.START_SEQ_NUM === "AUTO") {
     utility.getCurrentSeqNum(function(num) {
-        getMatches(num);
+        utility.scanApi(num);
     });
 }
 else if (process.env.START_SEQ_NUM) {
-    getMatches(process.env.START_SEQ_NUM);
+    utility.scanApi(process.env.START_SEQ_NUM);
 }
 else {
     //start at highest id in db
-    matches.findOne({
+    db.matches.findOne({
         upload: {
             $exists: false
         }
@@ -602,14 +593,14 @@ else {
         }
     }, function(err, doc) {
         if (err) {
-            logger.info(err);
+            utility.logger.info(err);
         }
-        getMatches(doc ? doc.match_seq_num + 1 : 0);
+        utility.scanApi(doc ? doc.match_seq_num + 1 : 0);
     });
 }
 setInterval(function untrackPlayers() {
-    logger.info("[UNTRACK] Untracking users...");
-    players.update({
+    utility.logger.info("[UNTRACK] Untracking users...");
+    db.players.update({
         last_visited: {
             $lt: moment().subtract(process.env.UNTRACK_INTERVAL_DAYS || 3, 'days').toDate()
         }
@@ -621,170 +612,8 @@ setInterval(function untrackPlayers() {
         multi: true
     }, function(err, num) {
         if (err) {
-            logger.info(err);
+            utility.logger.info(err);
         }
-        logger.info("[UNTRACK] Untracked %s users", num);
+        utility.logger.info("[UNTRACK] Untracked %s users", num);
     });
 }, 60 * 60 * 1000); //check every hour
-
-var server = app.listen(process.env.PORT || 3000, function() {
-    var host = server.address().address;
-    var port = server.address().port;
-    console.log('[WEB] listening at http://%s:%s', host, port);
-});
-
-function getMatches(seq_num) {
-    players.find({
-        track: 1
-    }, function(err, docs) {
-        if (err) {
-            return getMatches(seq_num);
-        }
-        //rebuild set of tracked players before every check
-        trackedPlayers = {};
-        docs.forEach(function(player) {
-            trackedPlayers[player.account_id] = true;
-        });
-        var url = utility.api_url + "/GetMatchHistoryBySequenceNum/V001/?key=" + process.env.STEAM_API_KEY + "&start_at_match_seq_num=" + seq_num;
-        utility.getData(url, function(err, data) {
-            if (err) {
-                return getMatches(seq_num);
-            }
-            var resp = data.result.matches;
-            logger.info("[API] seq_num: %s, found %s matches", seq_num, resp.length);
-            async.mapSeries(resp, insertMatch, function(err) {
-                if (err) {
-                    return getMatches(seq_num);
-                }
-                if (resp.length > 0) {
-                    seq_num = resp[resp.length - 1].match_seq_num + 1;
-                }
-                return getMatches(seq_num);
-            });
-        });
-    });
-}
-
-function processApiReq(job, cb) {
-    //process an api request
-    var payload = job.data.payload;
-    if (!job.data.url) {
-        logger.info(job);
-        cb("no url");
-    }
-    utility.getData(job.data.url, function(err, data) {
-        if (err) {
-            return cb(err);
-        }
-        if (data.response) {
-            //summaries response
-            async.map(data.response.players, insertPlayer, function(err) {
-                cb(err);
-            });
-        }
-        else if (payload.match_id) {
-            //response for single match details
-            var match = data.result;
-            insertMatch(match, function(err) {
-                cb(err);
-            });
-        }
-        else if (payload.account_id) {
-            //response for match history for single player
-            var resp = data.result.matches;
-            async.map(resp, function(match, cb2) {
-                utility.queueReq("api_details", match, function(err) {
-                    cb2(err);
-                });
-            }, function(err) {
-                cb(err);
-            });
-        }
-    });
-}
-
-function insertMatch(match, cb) {
-    var track = match.players.some(function(element) {
-        return (element.account_id in trackedPlayers);
-    });
-    //queued or untracked
-    match.parse_status = (track ? 0 : 3);
-    if (track || match.upload) {
-        var summaries = {
-            summaries_id: new Date(),
-            players: match.players
-        };
-        //queue for player names
-        utility.queueReq("api_summaries", summaries, function(err) {
-            if (err) return logger.info(err);
-        });
-        //parse if unparsed
-        if (match.parsed_data) {
-            match.parse_status = 2;
-        }
-        else {
-            utility.queueReq("parse", match, function(err) {
-                if (err) return logger.info(err);
-            });
-        }
-        matches.update({
-                match_id: match.match_id
-            }, {
-                $set: match
-            }, {
-                upsert: true
-            },
-            function(err) {
-                cb(err);
-            });
-    }
-    else {
-        cb(null);
-    }
-}
-
-function insertPlayer(player, cb) {
-    var account_id = Number(utility.convert64to32(player.steamid));
-    player.last_summaries_update = new Date();
-    players.update({
-        account_id: account_id
-    }, {
-        $set: player
-    }, {
-        upsert: true
-    }, function(err) {
-        cb(err);
-    });
-}
-
-function processUpload(job, cb) {
-    var fileName = job.data.payload.fileName;
-    utility.runParse(fileName, function(code, output) {
-        fs.unlink(fileName, function(err) {
-            logger.info(err);
-        });
-        if (!code) {
-            var api_container = utility.generateJob("api_details", {
-                match_id: output.match_id
-            });
-            //check api to fill rest of match info
-            utility.getData(api_container.url, function(err, body) {
-                if (err) {
-                    return cb(err);
-                }
-                var match = body.result;
-                match.parsed_data = output;
-                match.upload = true;
-                insertMatch(match, function(err) {
-                    cb(err);
-                });
-            });
-        }
-        else {
-            //only try once, mark done regardless of result
-            cb(null);
-        }
-    });
-}
-
-module.exports.app = app;
